Metric: Activity Frequency

Provides summary statistics about the frequency of activity types at the level of log, traces, cases, activity types.

activity_frequency(eventlog, level, append, ...)

# S3 method for eventlog
activity_frequency(eventlog, level = c("log", "trace",
  "activity", "case"), append = F, ...)

# S3 method for grouped_eventlog
activity_frequency(eventlog, level = c("log",
  "trace", "activity", "case"), append = F, ...)

Arguments

eventlog
The dataset to be used. Should be a (grouped) eventlog object. eventlog.

level
Level of granularity for the analysis: log, trace, case, activity. For more information, see vignette("metrics", "edeaR")

append
Logical, indicating whether to append results to original event log. Ignored when level is log or trace.

...
Deprecated arguments

Details

At log level, This metric shows the summary statistics of the frequency of activities throughout the complete event log.

On the level of the cases, this metric showsthe absolute and relative number of times the different activity types occur in each case. The absolute number shows the number of distinct activity types that occur in each of the cases. The relative number is calculated based on the total activity executions in the case.

On trace level, this metric presents the absolute and relative number of times a specific activity type occurs in each trace.

On the level of the activities, this metric provides the absolute and relative frequency of a specific activity in the complete event log.

Methods (by class)

eventlog: Compute activity frequency for eventlog

grouped_eventlog: Compute activity frequency for grouped event log

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.