Metric: Activity Presence

Calculates for each activity type in what percentage of cases it is present.

Usage

activity_presence(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for class 'eventlog'
activity_presence(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for class 'grouped_eventlog'
activity_presence(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for class 'activitylog'
activity_presence(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for class 'grouped_activitylog'
activity_presence(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

append

logical (default FALSE) : The arguments append and append_column have been deprecated in favour of augment.
Indicating whether to append results to original log. Ignored when level is "log" or "trace".

append_column

The arguments append and append_column have been deprecated in favour of augment.
Which of the output columns to append to log, if append = TRUE. Default column depends on chosen level.

sort

logical (default TRUE): Sort output on count. Only for levels with frequency count output.

eventlog

; please use log instead.

Details

An indication of variance can be the presence of the activities in the different cases. This metric shows for each activity the absolute number of cases in which each activity occurs together with its relative presence.

Methods (by class)

activity_presence(eventlog): Compute activity presence for an eventlog.

activity_presence(grouped_eventlog): Compute activity presence for a grouped_eventlog.

activity_presence(activitylog): Compute activity presence for an activitylog.

activity_presence(grouped_activitylog): Compute activity presence for a grouped_activitylog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also

Examples

if (FALSE) { # \dontrun{
data <- data.frame(case = rep("A",5),
activity_id = c("A","B","C","D","E"),
activity_instance_id = 1:5,
lifecycle_id = rep("complete",5),
timestamp = 1:5,
resource = rep("resource 1", 5))

log <- bupaR::eventlog(data,case_id = "case",
activity_id = "activity_id",
activity_instance_id = "activity_instance_id",
lifecycle_id = "lifecycle_id",
timestamp = "timestamp",
resource_id = "resource")

activity_presence(log)
} # }