Calculate queuing length

Usage

calculate_queuing_length(
  queueing_times,
  level = c("log", "activity", "resource"),
  time_interval
)

Arguments

queueing_times

Object of class queuing_times, returned by calculate_queuing_times.

level

character (default "log"): Level of granularity for the analysis: "log", "activity", "resource". For more information, see 'Details' below.

time_interval

The time interval after which the queue length should be calculated. For more information, see 'Details' below and the by argument of seq.Date.

Details

Argument level has the following options:

At log level, this metric calculates the total number of activity instances that are queued at a given moment in time.

At resource level, this metric calculates the total number activity instances that are queued for a given resource.

On activity level, this metric calculates the total number of activity instances that are queue for a given activity type.

Argument time_interval has the following options (see also the by argument of seq.Date):

A numeric as number of days.

An object of class difftime.

A character string, which could be one of "day", "week", "month", "quarter", or "year". The first day for which queue length is calculated, is the first timestamp found in the log.

See also