edeaR - Exploratory and Descriptive Event-based data Analysis in R

This package provides several useful techniques for Exploratory and Descriptive analysis of event based data in R, developed by the Business Informatics Research Group of Hasselt University.

See also

Author

Maintainer: Gert Janssenswillen gert.janssenswillen@uhasselt.be

Other contributors:

Gerard van Hulzen gerard.vanhulzen@uhasselt.be [contributor]

Marijke Swennen marijke.swennen@uhasselt.be [contributor]

Ivan Esin ivan.esin@student.uhasselt.be [contributor]

Hasselt University [copyright holder]