Metric: End activities

Analyse the end activities in the process.

end_activities(eventlog, level, append, ...)

# S3 method for eventlog
end_activities(
  eventlog,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = FALSE,
  append_column = NULL,
  sort = TRUE,
  ...
)

# S3 method for grouped_eventlog
end_activities(
  eventlog,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = FALSE,
  append_column = NULL,
  sort = TRUE,
  ...
)

Arguments

eventlog
The dataset to be used. Should be a (grouped) eventlog object. eventlog.

level
Level of granularity for the analysis: log, case, activity, resource or resource-activity. For more information, see vignette("metrics", "edeaR")

append
Logical, indicating whether to append results to original event log. Ignored when level is log or trace.

...
Deprecated arguments

append_column
Which of the output columns to append to log, if append = T. Default column depends on chosen level.

sort
Sort output on count. Defaults to TRUE. Only for levels with frequency count output.

Details

On log levels, this metric shows the absolute and relative number of activities that are the last activity in one or more of the cases.

On the level of the specific cases, this metric provides an overview of the end activity of each case.

On the activity level This metric calculates for each activity the absolute and relative number of cases that end with this activity type. Similar to the start activities metric, the relative number is calculated as a portion of the number of cases, being the number of \"opportunities\" that an activity could be the end activity. The cumulative sum is added to have an insight in the number of activities that is required to cover a certain part of the total.

On the level of the distinct resources, an overview of which resources execute the last activity per case can be of interest for a company. Probably this person is responsible for the correct communication to the customer.

Finally, on the resource-activity level, this metric shows for each occurring resource-activity combination the absolute and relative number of times this resource executes this activity as an end activity in a case.

Methods (by class)

eventlog: Compute end activities for eventlog

grouped_eventlog: Compute end activities for grouped eventlog

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.