Filter Start and End Conditions

Filters cases where the first and/or last activity adhere to the specified conditions.

Usage

filter_endpoints_condition(
  log,
  start_condition = NULL,
  end_condition = NULL,
  reverse = FALSE,
  eventlog = deprecated()
)

# S3 method for eventlog
filter_endpoints_condition(
  log,
  start_condition = NULL,
  end_condition = NULL,
  reverse = FALSE,
  eventlog = deprecated()
)

# S3 method for grouped_log
filter_endpoints_condition(
  log,
  start_condition = NULL,
  end_condition = NULL,
  reverse = FALSE,
  eventlog = deprecated()
)

# S3 method for activitylog
filter_endpoints_condition(
  log,
  start_condition = NULL,
  end_condition = NULL,
  reverse = FALSE,
  eventlog = deprecated()
)

filter_endpoints_conditions(
  log,
  start_condition = NULL,
  end_condition = NULL,
  reverse = FALSE,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

start_condition, end_condition

A logical condition.

reverse

logical (default FALSE): Indicating whether the selection should be reversed.

eventlog

; please use log instead.

Value

When given an object of type log, it will return a filtered log. When given an object of type grouped_log, the filter will be applied in a stratified way (i.e. each separately for each group). The returned log will be grouped on the same variables as the original log.

Methods (by class)

eventlog: Filters cases for an eventlog.

grouped_log: Filters cases for a grouped_log.

activitylog: Filters cases for an activitylog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also