Filter Life Cycle

Filters the log based on the life cycle identifier.

filter_lifecycle(
  log,
  lifecycles,
  reverse = FALSE,
  lifecycle = deprecated(),
  eventlog = deprecated()
)

# S3 method for eventlog
filter_lifecycle(
  log,
  lifecycles,
  reverse = FALSE,
  lifecycle = deprecated(),
  eventlog = deprecated()
)

# S3 method for grouped_eventlog
filter_lifecycle(
  log,
  lifecycles,
  reverse = FALSE,
  lifecycle = deprecated(),
  eventlog = deprecated()
)

ifilter_lifecycle(eventlog)

Arguments

log

eventlog: Object of class eventlog or derivatives (grouped_eventlog).

lifecycles

character vector: A vector of life cycle identifiers.

reverse

logical (default FALSE): Indicating whether the selection should be reversed.

lifecycle

; please use lifecycles instead.

eventlog

; please use log instead.

Value

When given an object of type log, it will return a filtered log. When given an object of type grouped_log, the filter will be applied in a stratified way (i.e. each separately for each group). The returned log will be grouped on the same variables as the original log.

Methods (by class)

eventlog: Filters based on life cycle identifiers for an eventlog.

grouped_eventlog: Filters based on life cycle identifiers a grouped_eventlog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also