Filter: Trace length percentile

Filters cases on length, using a percentile threshold.

filter_trace_length(eventlog, interval, percentage, reverse, ...)

# S3 method for eventlog
filter_trace_length(
  eventlog,
  interval = NULL,
  percentage = NULL,
  reverse = FALSE,
  ...
)

# S3 method for grouped_eventlog
filter_trace_length(
  eventlog,
  interval = NULL,
  percentage = NULL,
  reverse = FALSE,
  ...
)

ifilter_trace_length(eventlog)

Arguments

eventlog
The dataset to be used. Should be a (grouped) eventlog object.

interval
An trace length interval (numeric vector of length 2) to be used for absolute. Half open interval can be created using NA.

percentage
A percentage p to be used for relative filtering.

reverse
Logical, indicating whether the selection should be reversed.

...
Deprecated arguments.

Value

When given an eventlog, it will return a filtered eventlog. When given a grouped eventlog, the filter will be applied in a stratified way (i.e. each separately for each group). The returned eventlog will be grouped on the same variables as the original event log.

Details

This filter can be used by using an interval or by using a percentage. The percentage will always start with the shortest cases first and stop including cases when the specified percentile is reached. On the other hand, an absolute interval can be defined instead to filter cases which have a length in this interval.

Methods (by class)

eventlog: Filter event log

grouped_eventlog: Filter grouped event log

See also