Number of Self-loops

Provides information statistics on the number of self-loops in a trace.

Activity instances of the same activity type that are executed more than once immediately after each other by the same resource are in a self-loop ("length-1-loop"). If an activity instance of the same activity type is executed 3 times after each other by the same resource, this is defined as a "size 2 self-loop".

number_of_selfloops(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for eventlog
number_of_selfloops(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for grouped_eventlog
number_of_selfloops(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for activitylog
number_of_selfloops(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for grouped_activitylog
number_of_selfloops(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

type

character (default "all"): The type of repetitions: "all" (default), "repeat", or "redo". For more information, see 'Details' below.

level

character (default "log"): Level of granularity for the analysis: "log" (default), "case", "activity", "resource", or "resource-activity". For more information, see vignette("metrics", "edeaR") and 'Details' below.

append

logical (default FALSE) : Indicating whether to append results to original log. Ignored when level is "log" or "trace".

append_column

Which of the output columns to append to log, if append = TRUE. Default column depends on chosen level.

sort

logical (default TRUE): Sort output on count. Only for levels with frequency count output.

eventlog

; please use log instead.

Details

Two types of self-loops are defined, which can be chosen using the type argument:

"repeat" self-loops are activity executions of the same activity type that are executed immediately following each other by the same resource.

"redo" self-loops are activity executions of the same activity type that are executed immediately following each other by a different resource.

Argument level has the following options:

At "log" level, the summary statistics of the number of self-loops within a trace can give a first insight in the amount of waste in a log. As stated earlier, each combination of two occurrences of the same activity executed by the same resource will be counted as one repeat self-loop of this activity.

On "case" level, an overview is provided of the absolute and relative number of repeat and redo self-loops in each case. To calculate the relative number, each (repeat or redo) self-loop is counted as 1 occurrence, and the other activity instances are also counted as 1.

On "activity" level, the absolute and relative number of self-loops per activity can be an indication for which activities are causing the most waste in the process.

On "resource" level, this metric can give insights into which resources needs to repeat their work most often within a case, or for which resource the work they did should be redone by another resource within the same case. This metric shows the absolute and relative number of both repeat and redo self-loops for each resource in the log.

On "resource-activity" level, this metric can be used to get an insight in which activities are the most crucial for which resources. This metric shows the absolute and relative number of both repeat and redo self-loops for each of the resource-activity combinations that occur in the log. Two different relative numbers are provided here, one from the resource perspective and one from the activity perspective. At the resource perspective, the denominator is the total number of executions by the resource under consideration. At the activity perspective, the denominator is the total number of occurrences of the activity under consideration.

Methods (by class)

eventlog: Computes the number of self-loops for an eventlog.

grouped_eventlog: Computes the number of self-loops for a grouped_eventlog.

activitylog: Computes the number of self-loops for an activitylog.

grouped_activitylog: Computes the number of self-loops for a grouped_activitylog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also