Resource Involvement

Calculates for each resource or resource-activity combination in what percentage of cases it is present.

Next to the resource_frequency, the involvement of resources in cases can be of interest to, e.g., decide how "indispensable" they are. This metric is provided on three levels of analysis, which are the cases, the resources, and the resource-activity combinations.

resource_involvement(
  log,
  level = c("case", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for log
resource_involvement(
  log,
  level = c("case", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for grouped_log
resource_involvement(
  log,
  level = c("case", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

level

character (default "case"): Level of granularity for the analysis: "case" (default), "resource", or "resource-activity". For more information, see vignette("metrics", "edeaR") and 'Details' below.

append

logical (default FALSE) : Indicating whether to append results to original log. Ignored when level is "log" or "trace".

append_column

Which of the output columns to append to log, if append = TRUE. Default column depends on chosen level.

sort

logical (default TRUE): Sort output on count. Only for levels with frequency count output.

eventlog

; please use log instead.

Details

Argument level has the following options:

On "case" level, the absolute and relative number of distinct resources executing activities in each case is calculated, to get an overview of which cases are handled by a small amount of resources and which cases need more resources, indicating a higher level of variance in the process.

On "resource" level, this metric provides the absolute and relative number of cases in which each resource is involved, indicating which resources are more "necessary" within the process than the others.

On "resource-activity" level, this metric provides a list of all resource-activity combinations with the absolute and relative number of cases in which each resource-activity combination is involved.

Methods (by class)

log: Computes the resource involvement for a log.

grouped_log: Computes the resource involvement for a grouped_log.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also