Metric: Size of repetitions

Provides summary statistics on the sizes of repetitions.

Usage

size_of_repetitions(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  eventlog = deprecated()
)

# S3 method for class 'eventlog'
size_of_repetitions(
  log,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  eventlog = deprecated()
)

# S3 method for class 'grouped_eventlog'
size_of_repetitions(
  log,
  type = c("repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  eventlog = deprecated()
)

# S3 method for class 'activitylog'
size_of_repetitions(
  log,
  type = c("repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  eventlog = deprecated()
)

# S3 method for class 'grouped_activitylog'
size_of_repetitions(
  log,
  type = c("repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

type

character (default "all"): The type of repetitions: "all" (default), "repeat", or "redo". For more information, see 'Details' below.

level

character (default "log"): Level of granularity for the analysis: "log" (default), "case", "activity", "resource", or "resource-activity". For more information, see vignette("metrics", "edeaR") and 'Details' below.

append

logical (default FALSE) : The arguments append and append_column have been deprecated in favour of augment.
Indicating whether to append results to original log. Ignored when level is "log" or "trace".

append_column

The arguments append and append_column have been deprecated in favour of augment.
Which of the output columns to append to log, if append = TRUE. Default column depends on chosen level.

eventlog

; please use log instead.

Methods (by class)

size_of_repetitions(eventlog): Size of repetitions for eventlog

size_of_repetitions(grouped_eventlog): Size of repetitions for grouped event log

size_of_repetitions(activitylog): Size of repetitions for activitylog

size_of_repetitions(grouped_activitylog): Size of repetitions for grouped activitylog

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also