Start Activities

Analyse the start activities in the process.

Usage

start_activities(
  log,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for eventlog
start_activities(
  log,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for grouped_eventlog
start_activities(
  log,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for activitylog
start_activities(
  log,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

# S3 method for grouped_activitylog
start_activities(
  log,
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

level

character (default "log"): Level of granularity for the analysis: "log" (default), "case", "activity", "resource", or "resource-activity". For more information, see vignette("metrics", "edeaR") and 'Details' below.

append

logical (default FALSE) : The arguments append and append_column have been deprecated in favour of augment.
Indicating whether to append results to original log. Ignored when level is "log" or "trace".

append_column

The arguments append and append_column have been deprecated in favour of augment.
Which of the output columns to append to log, if append = TRUE. Default column depends on chosen level.

sort

logical (default TRUE): Sort output on count. Only for levels with frequency count output.

eventlog

; please use log instead.

Details

Argument level has the following options:

On "log" level, this metric shows the absolute and relative number of activities that are the first activity in one or more of the cases.

On "case" level, this metric provides an overview of the start activity of each case.

On "activity" level, this metric calculates for each activity the absolute and relative number of cases that start with this activity type. Similar to the end_activities metric, the relative number is calculated as a portion of the number of cases, being the number of "opportunities" that an activity could be the start activity. The cumulative sum is added to have an insight in the number of activities that is required to cover a certain part of the total.

On "resource" level, an overview of which resources execute the first activity per case are provided.

On "resource-activity" level, this metric shows for each occurring resource-activity combination the absolute and relative number of times this resource executes this activity as an start activity in a case.

Methods (by class)

start_activities(eventlog): Computes the start activities for an eventlog.

start_activities(grouped_eventlog): Computes the start activities for a grouped_eventlog.

start_activities(activitylog): Computes the start activities for an activitylog.

start_activities(grouped_activitylog): Computes the start activities for a grouped_activitylog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also