Throughput Time of Cases

Provides summary statistics concerning the throughput times of cases.

throughput_time(
  log,
  level = c("log", "trace", "case"),
  append = deprecated(),
  append_column = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  sort = TRUE,
  work_schedule = NULL,
  eventlog = deprecated()
)

# S3 method for eventlog
throughput_time(
  log,
  level = c("log", "trace", "case"),
  append = deprecated(),
  append_column = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  sort = TRUE,
  work_schedule = NULL,
  eventlog = deprecated()
)

# S3 method for grouped_eventlog
throughput_time(
  log,
  level = c("log", "trace", "case"),
  append = deprecated(),
  append_column = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  sort = TRUE,
  work_schedule = NULL,
  eventlog = deprecated()
)

# S3 method for activitylog
throughput_time(
  log,
  level = c("log", "trace", "case"),
  append = deprecated(),
  append_column = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  sort = TRUE,
  work_schedule = NULL,
  eventlog = deprecated()
)

# S3 method for grouped_activitylog
throughput_time(
  log,
  level = c("log", "trace", "case"),
  append = deprecated(),
  append_column = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  sort = TRUE,
  work_schedule = NULL,
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

level

character (default "log"): Level of granularity for the analysis: "log" (default), "trace", or "case". For more information, see vignette("metrics", "edeaR") and 'Details' below.

append

logical (default FALSE) : Indicating whether to append results to original log. Ignored when level is "log" or "trace".

append_column

Which of the output columns to append to log, if append = TRUE. Default column depends on chosen level.

units

character (default "auto"): The time unit in which the throughput times should be reported. Should be one of the following values: "auto" (default), "secs", "mins", "hours", "days", "weeks". See also the units argument of difftime.

sort

logical (default TRUE): Sort output on count. Only for levels with frequency count output.

work_schedule

A schedule of working hours. If provided, only working hours are counted as processing time.

eventlog

; please use log instead.

Details

Argument level has the following options:

At "log" level, the summary statistics describing the throughput time of cases in an aggregated fashion.

On "trace" level, the throughput time of the different process variants or traces in the log are calculated.

On "case" level, the throughput time is defined as the total duration of the case, or the difference between the timestamp of the end event and the timestamp of the start event of the case. Possible idle_time is also included in this calculation.

For other levels (e.g. "activity", "resource", or "resource-activity"), the throughput time is equal to the processing_time and are, therefore, not supported by this method.

Methods (by class)

eventlog: Computes throughput time for an eventlog.

grouped_eventlog: Computes throughput time for a grouped_eventlog.

activitylog: Computes throughput time for an activitylog.

grouped_activitylog: Computes throughput time for a grouped_activitylog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also