Calculate queuing times

Usage

calculate_queuing_times(
  log,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  eventlog = deprecated()
)

# S3 method for eventlog
calculate_queuing_times(
  log,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  eventlog = deprecated()
)

# S3 method for activitylog
calculate_queuing_times(
  log,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

units

character (default "auto"): The time unit in which the throughput times should be reported. Should be one of the following values: "auto" (default), "secs", "mins", "hours", "days", "weeks". See also the units argument of difftime.

eventlog

; please use log instead.

Value

Returns a list of all the activity instances, with the time they started, and the time since they were queued. Notice that this does not take into account any process model notion! The time since they are queued is the completion time of the previous activity in the log.

Methods (by class)

calculate_queuing_times(eventlog): Calculate queueing times for eventlog and grouped_eventlog.

calculate_queuing_times(activitylog): Calculate queueing times for activitylog and grouped_activitylog.

References

Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.

See also